Test Stack Smashing

+++ berkelib/utils_.h
@@ -15,7 +15,7 @@ void        str_ptrlen              (size_t *output, char *str);
 void 	    str_split				(char *** output, char *str, const char delimiter, size_t *count_ptr);
 void 		sleep_ms				(const uint32_t milliseconds);
 char * 		run_command				(char *command);
-uint16_t 	leftTrimTill			(char *strptr, const char ch);
+size_t 	    leftTrimTill			(char *strptr, const char ch);
 void 		getColumn				(char **output, char *str, const uint16_t column_no, REQUIRE_WITH_SIZE(const char *, delim));
 void 		readSearchGetFirstLine	(char **output, const char *path, REQUIRE_WITH_SIZE(const char *, search_key), const uint16_t search_column, REQUIRE_WITH_SIZE(const char*, delim));
